MIRTG-package(MIRTG)	R Documentation

Two-dimensional Compensatory Item Response Theory Model Graphing Package

Description

This package provides graphical representations of items, tests, and examinees using a compensatory MIRT model for two dimensions.

Details

Package: 	 MIRTG

Type: 	 Package

Version: 	 1.0

Date: 	 2010-03-14

License: 	 GPL version 1 or newer

The ¡°MIRT¡± package uses the "MASS" and "rgl" packages. The functions available in the ¡°MIRT¡± package are listed below. The functions in this package include: "vector.plot" "surface.2D" "surface.3D" "contour.plot" "difsurface.2D" "difsurface.3D" "centroid.plot" "clamshell.plot" "direction.plot" "correlation.plot" "inf.2D" "inf.3D" "Angle"

Author(s)

Zhan Shu, Terry Ackerman, Matthew Burke
Maintainer: zhan shu <z_shu@uncg.edu>

References

Ackerman, T. (1994), Using Multidimensional Item Response Theory to Understand What Items and Tests Are Measuring. Applied Measurement in Education,7(4),255-278.
Ackerman, T. (1996), Graphical Representation of Multidimensional Item Response Theory Analyses. Applied Psychological Measurement, 20(4), 311-329.
Ackerman.T, Gierl.M, (2003), Using Multidimensional Item Response Theory to Evaluate Educational and Psychological Tests. Educational Measurement: Issues and Practices.
Reckase.M, McKinley.R, (1983), The Definition of Difficulty and Discrimination of Multi-dimensional Item Response Theory Model. AERA, Montreal, April.
Reckase.M,(1985). The Difficulty of Test Item that Measure more than one Ability. AERR, Chicago, April.
Reckase.M, McKinley.R,(1991), The Discriminating Power of Items That Measure More Than One Dimension. Applied Psychological Measurement, 15(4), 361-373.

[Package MIRTG version 1.0 Index]