inf.2D(MIRTG)	R Documentation

Function to plot information surface in two-dimensional style

Description

The function produces a plot of the information surface at any user specified level of angular direction for the theta1, theta2 composite.

Usage

inf.2D(alpha, a1, a2, d, correlation = 0, azimuthal_angle = 0, colatitude_angle = 15, filename = "NA", type = "jpeg")

Arguments

alpha 	 The direction of information (the angle between the theta 1 and the information direction)

a1 	 A vector of the item discrimination parameters for theta1

a2 	 A vector of the item discrimination parameters for theta2

d 	 A vector of the item difficulty parameters

correlation 	 The correlation between theta 1 and theta 2. If the correlation is not equal to zero, the a1 and a2 will be transformed to maintain the orthogonal representation of the x and y axes.

azimuthal_angle 	 The azimuthal angle to view the plot (east-west rotation)

colatitude_angle 	 The colatitude angle to view the plot (north-south rotation)

filename 	 The user specified filename given to save the plot; if the filename is provided, the function will automatically save the plot for the users by the provided filename.

type 	 The format of files in which the user saves the plot. Options include (wmf, emf, png,jpg,jpeg,bmp,tif,tiff,ps,eps, pdf).

Details

The x-axis represents theta1, the y-axis represents theta2, and the z-axis represents the value of the amount of information. These plots are directly related to the clamshell plots, and are simply an alternative way to visualize the information available in the test/item. ¡°There is an information surface for each composite direction from 0 degrees to 90 degrees. The height of the surface, indicating the amount of information, at each of the 81 grid points for a 30 degree composite direction, for example, would be equal to the value indicated by the length of the vector in the 30 degree direction in the clamshell plot, at each trait level (Ackerman, 1994, p. 266).¡± To calculate the item/test information in a MIRT framework three key objects are required: 1), the positive response probability (the probability of a correct response,P; 2), the negative response probability (the probability of an incorrect response; 1 - P; and 3), the angular direction of the information (alpha). In the two dimensional case, the information formula is given as: MINF = P*(1- P)*(a1*cos(alpha1)+a2*cos(alpha2))^2, where alpha1+ alpha2=90 (Ackerman, Gierl, and Walker, 2003). Therefore, both of the thetas and the angular direction (alpha1 or alpha2) have an effect on the test information.

Author(s)

Zhan Shu, Terry Ackerman, Matthew Burke

References

Ackerman,T.(1996),Graphical Representation of Multidimensional Item Response Theory Analyses,Applied Psychological Measurement,20(4),311-329. Reckase.M, McKinley.R,(1991), The Discriminating Power of Items That Measure More Than One Dimension, Applied Psychological Measurement,15(4),361-373.

See Also

inf.3D

Examples

a1<-c(0.48,1.16,1.48,0.44,0.36,1.78,0.64,1.10,0.76,0.52,0.83,0.88,0.34,0.74,0.66)
a2<-c( 0.54,0.35,0.44,1.72,0.69,0.47,1.21,1.74,0.89,0.53,0.41,0.98,0.59,0.59,0.70)
d<-c( -1.11,0.29,1.51,-0.82,-1.89,-0.49,1.35,0.82,-0.21,-0.04,-0.68,0.22,-0.86,-1.33,1.21)
inf.2D(pi/3, a1, a2, d)
inf.2D(pi/2,a1,a2,d)# check the information at different direction
inf.2D(pi/2,a1,a2,d,azimuthal_angle=15) # change the view angle
inf.2D(pi/2,a1,a2,d,colatitude_angle=0) # change the view angle

[Package MIRTG version 1.0 Index]